bug fixing

+++ src/database/schemas/schema.ts
@@ -151,10 +151,10 @@ export const accounts = pgTable("accounts", {
   placeId: text("place_id"),
 
   // Budget tracking
-  weeklyBudget: numeric("weekly_budget", { precision: 10, scale: 2 }).notNull(),
-  dailyBudget: numeric("daily_budget", { precision: 10, scale: 2 }),
-  monthlyBudget: numeric("monthly_budget", { precision: 10, scale: 2 }),
-  annualBudget: numeric("annual_budget", { precision: 10, scale: 2 }),
+  weeklyBudget: numeric("weekly_budget", { precision: 16, scale: 2 }).notNull(),
+  dailyBudget: numeric("daily_budget", { precision: 16, scale: 2 }),
+  monthlyBudget: numeric("monthly_budget", { precision: 16, scale: 2 }),
+  annualBudget: numeric("annual_budget", { precision: 16, scale: 2 }),
 
   currentAllocation: budgetAllocationEnum("current_allocation")
     .notNull()
